Refactor ui into modules
