--- js/ui.js
@@ -1,637 +1,0 @@
-// ui.js
-import { TokenType, tokenizeForHighlighting } from './tokenizer.js';
-import { Parser } from './parser.js'; // Needed for the run button
-import { tokenizeForParser } from './tokenizer.js'; // Needed for the run button
-
-let currentEditorHighlightLine = null;
-let uiInterpreterInstance = null; // To store the interpreter instance
-let suppressTabScroll = false; // Prevent scroll when selecting tab programmatically
-
-function getCursorLineNumber() {
-    if (!elements.inputArea) return null;
-    const value = elements.inputArea.value.slice(0, elements.inputArea.selectionStart);
-    return value.split(/\r?\n/).length;
-}
-
-// DOM Elements Cache
-const elements = {};
-
-function queryElements() {
-    elements.inputArea = document.getElementById('pipeDataInput');
-    elements.highlightingOverlay = document.getElementById('highlightingOverlay');
-    elements.lineNumbers = document.getElementById('lineNumbers');
-    elements.varBlockIndicator = document.getElementById('varBlockIndicator');
-    elements.astOutputArea = document.getElementById('astOutput');
-    elements.logOutputEl = document.getElementById('logOutput');
-    elements.peekTabsContainerEl = document.getElementById('peekTabsContainer');
-    elements.peekOutputsDisplayAreaEl = document.getElementById('peekOutputsDisplayArea');
-    elements.runButton = document.getElementById('runButton');
-    elements.clearButton = document.getElementById('clearButton');
-    elements.openFileButton = document.getElementById('openScriptFileButton');
-    elements.saveFileButton = document.getElementById('saveScriptFileButton');
-    elements.csvFileInputEl = document.getElementById('csvFileInput');
-    elements.fileInputContainerEl = document.getElementById('fileInputContainer');
-    elements.filePromptMessageEl = document.getElementById('filePromptMessage');
-    elements.exportPeekButton = document.getElementById('exportPeekButton'); // <-- New element
-}
-
-
-function escapeHtml(unsafe) {
-    if (typeof unsafe !== 'string') {
-        unsafe = String(unsafe);
-    }
-    return unsafe
-         .replace(/&/g, "&amp;")
-         .replace(/</g, "&lt;")
-         .replace(/>/g, "&gt;")
-         .replace(/"/g, "&quot;")
-         .replace(/'/g, "&#039;");
-}
-
-function updateLineNumbers() {
-    if (!elements.inputArea || !elements.lineNumbers) return;
-    const lineCount = elements.inputArea.value.split(/\r?\n/).length || 1;
-    let numbers = '';
-    for (let i = 1; i <= Math.min(lineCount, 999); i++) {
-        numbers += String(i).padStart(3, ' ') + '\n';
-    }
-    elements.lineNumbers.textContent = numbers;
-    elements.lineNumbers.scrollTop = elements.inputArea.scrollTop;
-}
-
-export function applySyntaxHighlighting(text, activePeekLine = null) {
-    const tokens = tokenizeForHighlighting(text);
-    let html = '';
-    let currentVarBlockStyleIndex = 0;
-    const varBlockStyles = ['var-block-bg-1', 'var-block-bg-2', 'var-block-bg-3', 'var-block-bg-4'];
-    let inVarBlock = false;
-    let blockContentHtml = '';
-    const activePeekHighlightID = "active-editor-peek-highlight";
-
-    function closeCurrentVarBlock() {
-        if (inVarBlock && blockContentHtml.trim() !== '') {
-            html += `<div class="var-block ${varBlockStyles[currentVarBlockStyleIndex % varBlockStyles.length]}">${blockContentHtml}</div>`;
-            currentVarBlockStyleIndex++;
-        } else if (blockContentHtml.trim() !== '') {
-            html += blockContentHtml;
-        }
-        blockContentHtml = '';
-        inVarBlock = false;
-    }
-
-    const firstTokenFound = new Set();
-
-    for (let i = 0; i < tokens.length; i++) {
-        const token = tokens[i];
-        let tokenHtml = '';
-        const escapedValue = escapeHtml(token.value);
-
-        const isFirstNonWhitespace =
-            !firstTokenFound.has(token.line) &&
-            token.type !== TokenType.WHITESPACE &&
-            token.type !== TokenType.NEWLINE;
-        if (isFirstNonWhitespace) firstTokenFound.add(token.line);
-
-        if (token.type === TokenType.KEYWORD && token.value.toUpperCase() === 'VAR') {
-            closeCurrentVarBlock();
-            inVarBlock = true;
-        }
-
-        let classes = '';
-        let idAttribute = '';
-
-        const shouldHighlight = token.line === activePeekLine && isFirstNonWhitespace;
-
-        switch (token.type) {
-            case TokenType.KEYWORD:
-                classes = 'token-keyword';
-                break;
-            case TokenType.STRING_LITERAL:
-                classes = 'token-string_literal';
-                break;
-            case TokenType.NUMBER_LITERAL:
-                classes = 'token-number_literal';
-                break;
-            case TokenType.COMMENT:
-                classes = 'token-comment';
-                break;
-            case TokenType.OPERATOR:
-                classes = 'token-operator';
-                break;
-            case TokenType.IDENTIFIER:
-                classes = 'token-identifier';
-                break;
-            case TokenType.PUNCTUATION:
-                classes = 'token-punctuation';
-                break;
-            case TokenType.NEWLINE: tokenHtml = '\n'; break;
-            case TokenType.WHITESPACE: tokenHtml = escapedValue; break;
-            default: tokenHtml = escapedValue;
-        }
-
-        if (!tokenHtml) {
-            if (shouldHighlight) {
-                classes += ' active-peek-line-highlight';
-                idAttribute = ` id="${activePeekHighlightID}"`;
-            }
-            tokenHtml = `<span class="${classes}"${idAttribute}>${escapedValue}</span>`;
-        }
-
-        if (inVarBlock) {
-            blockContentHtml += tokenHtml;
-        } else {
-            html += tokenHtml;
-        }
-    }
-    closeCurrentVarBlock();
-    return html + '\n\n'; // Extra newlines for padding if needed
-}
-
-
-function generatePeekHtmlForDisplay(datasetToPeek, varName, line) {
-    let outputHTML = `<h3 class="text-md font-semibold mb-2 text-gray-100">Data for: VAR "${varName || 'Current Context'}" (PEEK at Line ${line})</h3>`;
-
-    if (!datasetToPeek) {
-        outputHTML += '<p class="text-gray-400">No dataset loaded to PEEK.</p>';
-        return outputHTML;
-    }
-
-    if (!Array.isArray(datasetToPeek)) {
-        outputHTML += `<p class="text-gray-400">PEEKed data is not tabular (Type: ${typeof datasetToPeek}). Preview may be limited.</p>`;
-        if (typeof datasetToPeek === 'object') {
-            outputHTML += `<pre class="text-gray-400">${escapeHtml(JSON.stringify(datasetToPeek, null, 2))}</pre>`;
-        } else {
-            outputHTML += `<pre class="text-gray-400">${escapeHtml(String(datasetToPeek))}</pre>`;
-        }
-        return outputHTML;
-    }
-
-    if (datasetToPeek.length > 0 && typeof datasetToPeek[0] === 'object') {
-        const peekRowCount = 10;
-        const dataToDisplay = datasetToPeek.slice(0, peekRowCount);
-        const headers = Object.keys(dataToDisplay[0]);
-            let tableHtml = '<table><thead><tr>';
-            headers.forEach(header => tableHtml += `<th>${escapeHtml(String(header))}</th>`);
-            tableHtml += '</tr></thead><tbody>';
-            dataToDisplay.forEach(row => {
-                tableHtml += '<tr>';
-                headers.forEach(header => {
-                    const value = row[header];
-                    tableHtml += `<td>${value === null || value === undefined ? '' : escapeHtml(String(value))}</td>`;
-                });
-                tableHtml += '</tr>';
-            });
-            tableHtml += `</tbody></table>`;
-             if (datasetToPeek.length > peekRowCount) {
-                tableHtml += `<p class="text-xs text-gray-400 mt-2">Showing first ${peekRowCount} of ${datasetToPeek.length} rows (basic preview).</p>`;
-            } else {
-                tableHtml += `<p class="text-xs text-gray-400 mt-2">Showing all ${datasetToPeek.length} rows (basic preview).</p>`;
-            }
-            outputHTML += tableHtml;
-        if (datasetToPeek.length > peekRowCount) {
-            outputHTML += `<p class="text-xs text-gray-400 mt-2">Showing first ${peekRowCount} of ${datasetToPeek.length} rows.</p>`;
-        } else {
-            outputHTML += `<p class="text-xs text-gray-400 mt-2">Showing all ${datasetToPeek.length} rows.</p>`;
-        }
-        return outputHTML;
-    }
-
-    outputHTML += `<pre class="text-gray-400">${escapeHtml(JSON.stringify(datasetToPeek, null, 2))}</pre>`;
-    return outputHTML;
-}
-
-function renderPeekOutputsUI() {
-    if (!elements.peekTabsContainerEl || !elements.peekOutputsDisplayAreaEl || !uiInterpreterInstance) {
-        console.error("Peek UI elements not found or interpreter not set!");
-        return;
-    }
-
-    elements.peekTabsContainerEl.innerHTML = '';
-    elements.peekOutputsDisplayAreaEl.innerHTML = '';
-    
-    const peekOutputs = uiInterpreterInstance.peekOutputs;
-    const stepOutputs = uiInterpreterInstance.stepOutputs;
-
-    if (peekOutputs.length === 0 && stepOutputs.length === 0) {
-        elements.peekOutputsDisplayAreaEl.innerHTML = '<div class="output-box-placeholder">No PEEK outputs to display.</div>';
-        if (elements.exportPeekButton) elements.exportPeekButton.classList.add('hidden');
-        clearEditorPeekHighlight();
-        return;
-    }
-
-    if (elements.exportPeekButton) elements.exportPeekButton.classList.remove('hidden');
-
-    stepOutputs.forEach((stepData, index) => {
-        const tabButton = document.createElement('button');
-        tabButton.classList.add('peek-tab');
-        tabButton.textContent = stepData.varName;
-        tabButton.dataset.target = stepData.id;
-        tabButton.dataset.stepIndex = index;
-        tabButton.dataset.line = stepData.line;
-
-        const contentDiv = document.createElement('div');
-        contentDiv.id = stepData.id;
-        contentDiv.classList.add('peek-content');
-        contentDiv.innerHTML = generatePeekHtmlForDisplay(stepData.dataset, stepData.varName, stepData.line);
-
-        elements.peekTabsContainerEl.appendChild(tabButton);
-        elements.peekOutputsDisplayAreaEl.appendChild(contentDiv);
-
-        tabButton.addEventListener('click', () => {
-            elements.peekTabsContainerEl.querySelectorAll('.peek-tab').forEach(tab => tab.classList.remove('active-peek-tab'));
-            elements.peekOutputsDisplayAreaEl.querySelectorAll('.peek-content').forEach(content => content.classList.remove('active-peek-content'));
-
-            tabButton.classList.add('active-peek-tab');
-            contentDiv.classList.add('active-peek-content');
-
-            if (elements.inputArea && elements.highlightingOverlay) {
-                elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(elements.inputArea.value, stepData.line);
-                currentEditorHighlightLine = stepData.line;
-                const highlightedSpan = elements.highlightingOverlay.querySelector('#active-editor-peek-highlight');
-                if (highlightedSpan) {
-                    const scrollTargetOffset = highlightedSpan.offsetTop;
-                    const inputAreaVisibleHeight = elements.inputArea.clientHeight;
-                    const desiredScrollTop = scrollTargetOffset - (inputAreaVisibleHeight / 3);
-
-                    if (!suppressTabScroll) {
-                        elements.inputArea.scrollTop = Math.max(0, desiredScrollTop);
-                        elements.highlightingOverlay.scrollTop = elements.inputArea.scrollTop;
-                    }
-                }
-                updateVarBlockIndicator(stepData.line);
-            }
-        });
-
-        if (index === 0 && peekOutputs.length === 0) {
-            tabButton.click();
-        }
-    });
-
-    peekOutputs.forEach((peekData, index) => {
-        const tabButton = document.createElement('button');
-        tabButton.classList.add('peek-tab');
-        tabButton.textContent = peekData.varName;
-        tabButton.dataset.target = peekData.id;
-        tabButton.dataset.peekIndex = index; // Store index for easy data retrieval
-        tabButton.dataset.line = peekData.line;
-
-        const contentDiv = document.createElement('div');
-        contentDiv.id = peekData.id;
-        contentDiv.classList.add('peek-content');
-        // Pass the raw dataset (array or other structure)
-        contentDiv.innerHTML = generatePeekHtmlForDisplay(peekData.dataset, peekData.varName, peekData.line);
-
-        elements.peekTabsContainerEl.appendChild(tabButton);
-        elements.peekOutputsDisplayAreaEl.appendChild(contentDiv);
-
-        tabButton.addEventListener('click', () => {
-            elements.peekTabsContainerEl.querySelectorAll('.peek-tab').forEach(tab => tab.classList.remove('active-peek-tab'));
-            elements.peekOutputsDisplayAreaEl.querySelectorAll('.peek-content').forEach(content => content.classList.remove('active-peek-content'));
-
-            tabButton.classList.add('active-peek-tab');
-            contentDiv.classList.add('active-peek-content');
-
-            if (elements.inputArea && elements.highlightingOverlay) {
-                elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(elements.inputArea.value, peekData.line);
-                currentEditorHighlightLine = peekData.line;
-                
-                const highlightedSpan = elements.highlightingOverlay.querySelector('#active-editor-peek-highlight');
-                if (highlightedSpan) {
-                    const scrollTargetOffset = highlightedSpan.offsetTop;
-                    const inputAreaVisibleHeight = elements.inputArea.clientHeight;
-                    const desiredScrollTop = scrollTargetOffset - (inputAreaVisibleHeight / 3);
-
-                    if (!suppressTabScroll) {
-                        elements.inputArea.scrollTop = Math.max(0, desiredScrollTop);
-                        elements.highlightingOverlay.scrollTop = elements.inputArea.scrollTop;
-                    }
-                }
-                updateVarBlockIndicator(peekData.line);
-            }
-        });
-
-        if (index === 0) { 
-            tabButton.click(); // Programmatically click the first tab to activate it and its logic
-        }
-    });
-}
-
-function clearEditorPeekHighlight() {
-    if (elements.inputArea && elements.highlightingOverlay && currentEditorHighlightLine !== null) {
-        elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(elements.inputArea.value, null);
-        currentEditorHighlightLine = null;
-    } else if (elements.inputArea && elements.highlightingOverlay && currentEditorHighlightLine === null) {
-         elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(elements.inputArea.value, null);
-    }
-}
-
-function showOutputForLine(lineNumber) {
-    if (!elements.peekTabsContainerEl) return;
-    const tabSelectors = [
-        `.peek-tab[data-peek-index][data-line='${lineNumber}']`,
-        `.peek-tab[data-step-index][data-line='${lineNumber}']`
-    ];
-    for (const sel of tabSelectors) {
-        const tab = elements.peekTabsContainerEl.querySelector(sel);
-        if (tab) {
-            suppressTabScroll = true;
-            tab.click();
-            suppressTabScroll = false;
-            return;
-        }
-    }
-}
-
-function updateVarBlockIndicator(lineNumber) {
-    if (!elements.varBlockIndicator || !elements.inputArea) return;
-    const lines = elements.inputArea.value.split(/\r?\n/);
-    if (lineNumber < 1 || lineNumber > lines.length) {
-        elements.varBlockIndicator.style.display = 'none';
-        return;
-    }
-
-    let start = null;
-    for (let i = lineNumber - 1; i >= 0; i--) {
-        if (/^\s*VAR\b/i.test(lines[i])) { start = i + 1; break; }
-    }
-    if (start === null) {
-        elements.varBlockIndicator.style.display = 'none';
-        return;
-    }
-    let end = lines.length;
-    for (let i = lineNumber; i < lines.length; i++) {
-        if (/^\s*VAR\b/i.test(lines[i])) { end = i; break; }
-        if (/^\s*THEN\s+PEEK\b/i.test(lines[i])) { end = i + 1; break; }
-    }
-
-    // Trim trailing empty lines so the indicator height matches syntax highlighting
-    while (end > start && /^\s*$/.test(lines[end - 1])) {
-        end--;
-    }
-
-    const style = getComputedStyle(elements.inputArea);
-    const lineHeight = parseFloat(style.lineHeight);
-    const paddingTop = parseFloat(style.paddingTop);
-    const paddingLeft = parseFloat(style.paddingLeft);
-    const borderLeft = parseFloat(style.borderLeftWidth);
-    const borderTop = parseFloat(style.borderTopWidth);
-
-    const top = paddingTop + borderTop + (start - 1) * lineHeight - elements.inputArea.scrollTop;
-    const height = (end - start + 1) * lineHeight;
-    const left = paddingLeft + borderLeft - elements.inputArea.scrollLeft - 6; // small gap from text
-
-    elements.varBlockIndicator.style.top = `${top}px`;
-    elements.varBlockIndicator.style.left = `${left}px`;
-    elements.varBlockIndicator.style.height = `${height}px`;
-    elements.varBlockIndicator.style.display = 'block';
-}
-
-
-function clearOutputs() {
-    if (uiInterpreterInstance) {
-         // Clear interpreter's log output element
-        if (elements.logOutputEl) elements.logOutputEl.innerHTML = 'Logs will appear here...<br>';
-        // Clear interpreter's peek and step data and related UI
-        uiInterpreterInstance.peekOutputs = [];
-        uiInterpreterInstance.stepOutputs = [];
-    }
-   
-    if (elements.peekTabsContainerEl) elements.peekTabsContainerEl.innerHTML = '';
-    if (elements.peekOutputsDisplayAreaEl) {
-        elements.peekOutputsDisplayAreaEl.innerHTML = '<div class="output-box-placeholder">Peek results will appear here when a script is run.</div>';
-    }
-    if (elements.astOutputArea) {
-        elements.astOutputArea.textContent = 'AST will appear here...';
-        elements.astOutputArea.classList.remove('error-box');
-    }
-    if (elements.fileInputContainerEl) {
-        elements.fileInputContainerEl.classList.add('hidden');
-    }
-    if (elements.exportPeekButton) elements.exportPeekButton.classList.add('hidden'); // <-- Hide export button
-    clearEditorPeekHighlight();
-    if (elements.varBlockIndicator) elements.varBlockIndicator.style.display = 'none';
-}
-
-// --- START NEW FUNCTION ---
-function handleExportPeek() {
-    if (!uiInterpreterInstance || !elements.peekTabsContainerEl) return;
-
-    const activeTab = elements.peekTabsContainerEl.querySelector('.peek-tab.active-peek-tab');
-    if (!activeTab) {
-        alert("No active PEEK tab found to export.");
-        return;
-    }
-
-    const peekIndex = parseInt(activeTab.dataset.peekIndex, 10);
-    const peekDataEntry = uiInterpreterInstance.peekOutputs[peekIndex];
-
-    if (!peekDataEntry || !peekDataEntry.dataset) {
-        alert("Could not find data for the active PEEK tab.");
-        return;
-    }
-
-    const dataset = peekDataEntry.dataset;
-
-    if (Array.isArray(dataset) && dataset.length > 0 && typeof dataset[0] === 'object') {
-        // Basic CSV conversion for array of objects
-        try {
-            const papaCSV = Papa.unparse(dataset);
-            const blob = new Blob([papaCSV], { type: 'text/csv;charset=utf-8;' });
-            const link = document.createElement("a");
-            const url = URL.createObjectURL(blob);
-            link.setAttribute("href", url);
-            link.setAttribute("download", `peek_export_${peekDataEntry.varName}_L${peekDataEntry.line}.csv`);
-            link.style.visibility = 'hidden';
-            document.body.appendChild(link);
-            link.click();
-            document.body.removeChild(link);
-            uiInterpreterInstance.log(`Exported PEEK data (array of objects) for VAR "${peekDataEntry.varName}" (Line ${peekDataEntry.line}) to CSV.`);
-        } catch (error) {
-            console.error("Error exporting array of objects to CSV:", error);
-            alert("Failed to export data to CSV. See console for details.");
-            uiInterpreterInstance.log(`Error exporting PEEK (array of objects) for VAR "${peekDataEntry.varName}" to CSV: ${error.message}`);
-        }
-    } else {
-        alert("The active PEEK tab data is not in a format that can be exported to CSV.");
-        uiInterpreterInstance.log(`PEEK data for VAR "${peekDataEntry.varName}" (Line ${peekDataEntry.line}) is not exportable to CSV (type: ${typeof dataset})`);
-    }
-}
-// --- END NEW FUNCTION ---
-
-
-async function saveScriptToFile() {
-    if (!elements.inputArea || typeof window.showSaveFilePicker !== 'function') return;
-    try {
-        const handle = await window.showSaveFilePicker({
-            types: [{ description: 'PipeData Script', accept: { 'text/plain': ['.pd'] } }]
-        });
-        const writable = await handle.createWritable();
-        await writable.write(elements.inputArea.value);
-        await writable.close();
-        if (uiInterpreterInstance) uiInterpreterInstance.log('Script saved to file.');
-    } catch (e) {
-        // User cancelled
-    }
-}
-
-async function loadScriptFromFile() {
-    if (!elements.inputArea || typeof window.showOpenFilePicker !== 'function') return;
-    try {
-        const [handle] = await window.showOpenFilePicker({
-            types: [{ description: 'PipeData Script', accept: { 'text/plain': ['.pd', '.txt'] } }]
-        });
-        const file = await handle.getFile();
-        const text = await file.text();
-        elements.inputArea.value = text;
-        updateLineNumbers();
-        if (elements.highlightingOverlay) {
-            elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(text, null);
-            currentEditorHighlightLine = null;
-        }
-        if (uiInterpreterInstance) uiInterpreterInstance.log('Loaded script from file.');
-    } catch (e) {
-        // User cancelled
-    }
-}
-
-
-const fallbackScript = `VAR "cities"
-THEN LOAD_CSV FILE "exampleCities.csv"
-THEN SELECT population, id
-THEN WITH COLUMN population_millions = population / 1000000
-THEN PEEK
-VAR "people"
-THEN LOAD_CSV FILE "examplePeople.csv"
-THEN PEEK
-THEN JOIN cities ON city_id=id TYPE "LEFT"
-THEN SELECT name, age, population, population_millions
-THEN FILTER name STARTSWITH "A"
-THEN PEEK`;
-
-async function loadDefaultScript() {
-    try {
-        const resp = await fetch('examples/default.pd');
-        if (resp.ok) {
-            return await resp.text();
-        }
-    } catch (e) {
-        // ignore and fall back
-    }
-    return fallbackScript;
-}
-
-export async function initUI(interpreter) {
-    uiInterpreterInstance = interpreter; // Store interpreter instance for UI use
-    queryElements(); // Get all DOM elements
-    updateLineNumbers();
-
-    const defaultScript = await loadDefaultScript();
-    if (elements.inputArea && elements.inputArea.value.trim() === '') {
-        elements.inputArea.value = defaultScript;
-        updateLineNumbers();
-        if (elements.highlightingOverlay) {
-            elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(defaultScript, null);
-        }
-        if (elements.varBlockIndicator) updateVarBlockIndicator(1);
-    }
-    clearOutputs(); // Initial clear
-
-    elements.inputArea?.addEventListener('input', () => {
-        const text = elements.inputArea.value;
-        updateLineNumbers();
-        elements.highlightingOverlay.innerHTML = applySyntaxHighlighting(text, null); // Clear peek highlight on input
-        currentEditorHighlightLine = null;
-        elements.highlightingOverlay.scrollTop = elements.inputArea.scrollTop;
-        elements.highlightingOverlay.scrollLeft = elements.inputArea.scrollLeft;
-        const line = getCursorLineNumber();
-        if (line) updateVarBlockIndicator(line);
-    });
-
-    elements.inputArea?.addEventListener('scroll', () => {
-        elements.highlightingOverlay.scrollTop = elements.inputArea.scrollTop;
-        elements.highlightingOverlay.scrollLeft = elements.inputArea.scrollLeft;
-        updateLineNumbers();
-        const line = getCursorLineNumber();
-        if (line) updateVarBlockIndicator(line);
-    });
-
-    if (elements.inputArea) {
-        new ResizeObserver(() => {
-            elements.highlightingOverlay.style.height = elements.inputArea.clientHeight + 'px';
-            elements.highlightingOverlay.style.width = elements.inputArea.clientWidth + 'px';
-            if (elements.lineNumbers) {
-                elements.lineNumbers.style.height = elements.inputArea.clientHeight + 'px';
-            }
-        }).observe(elements.inputArea);
-    }
-
-    elements.csvFileInputEl?.addEventListener('change', (event) => {
-        const file = event.target.files[0];
-        if (uiInterpreterInstance && uiInterpreterInstance.fileResolve) {
-            uiInterpreterInstance.fileResolve(file);
-            uiInterpreterInstance.fileResolve = null; // Consume the resolver
-        }
-    });
-
-    elements.runButton?.addEventListener('click', async () => {
-        if (!uiInterpreterInstance) return;
-
-        const script = elements.inputArea.value;
-        elements.astOutputArea.classList.remove('error-box');
-        elements.astOutputArea.textContent = 'Parsing...';
-        
-        if (elements.logOutputEl) elements.logOutputEl.innerHTML = 'Logs will appear here...<br>';
-        if (elements.peekTabsContainerEl) elements.peekTabsContainerEl.innerHTML = '';
-        if (elements.peekOutputsDisplayAreaEl) elements.peekOutputsDisplayAreaEl.innerHTML = '<div class="output-box-placeholder">Peek results will appear here when a script is run.</div>';
-        if (elements.exportPeekButton) elements.exportPeekButton.classList.add('hidden'); // <-- Hide export on run
-        clearEditorPeekHighlight();
-        uiInterpreterInstance.clearInternalState(); 
-
-        try {
-            const tokensForParser = tokenizeForParser(script);
-            const parser = new Parser(tokensForParser);
-            const ast = parser.parse();
-            elements.astOutputArea.textContent = JSON.stringify(ast, null, 2);
-
-            await uiInterpreterInstance.run(ast); 
-
-        } catch (e) {
-            elements.astOutputArea.classList.add('error-box');
-            const errorMessage = e instanceof Error ? e.message : (typeof e === 'string' ? e : JSON.stringify(e));
-            const stackTrace = e instanceof Error && e.stack ? "\n" + e.stack.split('\n').slice(0,3).join('\n') : '';
-            elements.astOutputArea.textContent = `Error: ${errorMessage}${stackTrace}`;
-            uiInterpreterInstance.log(`Error during parsing or execution: ${errorMessage}`);
-            console.error("Full error object:", e);
-        } finally {
-            renderPeekOutputsUI(); 
-        }
-    });
-
-    elements.clearButton?.addEventListener('click', () => {
-        clearOutputs();
-    });
-
-    elements.saveFileButton?.addEventListener('click', saveScriptToFile);
-    elements.openFileButton?.addEventListener('click', loadScriptFromFile);
-
-    // --- START NEW EVENT LISTENER ---
-    elements.exportPeekButton?.addEventListener('click', handleExportPeek);
-    // --- END NEW EVENT LISTENER ---
-
-    elements.inputArea?.addEventListener('keyup', () => {
-        const line = getCursorLineNumber();
-        if (line) {
-            showOutputForLine(line);
-            updateVarBlockIndicator(line);
-        }
-    });
-    elements.inputArea?.addEventListener('click', () => {
-        const line = getCursorLineNumber();
-        if (line) {
-            showOutputForLine(line);
-            updateVarBlockIndicator(line);
-        }
-    });
-}
-
-export { renderPeekOutputsUI, generatePeekHtmlForDisplay, clearOutputs };
